pyproject.toml
[project]
name = "ai-image-processor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.41.0",
    "google-generativeai>=0.8.5",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
